+++ cs4090/webApp/admin.py
@@ -1,3 +1,11 @@
 from django.contrib import admin
+from webApp import models
 
-# Register your models here.
+
+@admin.register(models.UserStreak)
+class StreakAdmin(admin.ModelAdmin):
+    model = models.UserStreak
+    list_display = [
+        "current_streak",
+        "last_activity_date"
+    ]
